trader detail: back button on top and cleaner view on mobile and desktop
