+++ src/app/pipes/pipes.modules.ts
@@ -1,8 +1,10 @@
 import { NgModule } from '@angular/core';
 import { LabelPipe } from './label.pipe';
 import { SafePipe } from './safe.pipe';
+import { Nl2BrPipeModule } from 'nl2br-pipe';
 
 @NgModule({
+  imports: [Nl2BrPipeModule],
   declarations: [LabelPipe, SafePipe],
   exports: [LabelPipe, SafePipe],
 })
