+++ src/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="de">
+<html lang="de-DE">
   <head>
     <meta charset="utf-8" />
     <title>lokalkauf - Dein Lieblingsladen und Restaurant Online!</title>
